pyproject.toml
[project]
name = "mlProject"
version = "0.0.1"
description = "A flexible end-to-end Machine Learning project"
readme = "README.md"
requires-python = ">=3.12,<3.14"


dependencies = [
    "pandas",
    "numpy",
    "matplotlib",
    "seaborn",
    "scipy",
    "scikit-learn",
    "python-box>=7.3.2",
    "pyyaml>=6.0.3",
    "ensure>=1.0.4",
    "joblib>=1.5.3",
    "ipykernel>=7.1.0",
    "sweetviz>=2.3.1",
    "setuptools>=80.9.0",
    "notebook>=7.5.1",
    "pip>=25.3",
    "statsmodels>=0.14.6",
    "dvc>=3.65.0",
    "mlflow>=3.8.1",
    "dagshub>=0.6.4",
    "wandb>=0.23.1",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pydantic>=2.12.5",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.21",
    "locust>=2.43.0",
    "evidently>=0.7.18",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp-proto-http>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "python-dotenv>=1.2.1",
    "opentelemetry-exporter-otlp>=1.39.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mlProject"]
